requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
streamlit>=1.40.0
python-dotenv>=1.0.0
